Clean up format APIs in LocalizationService.java @W-5458971 (#3428)
* Clean up format APIs in LocalizationService.java

@bug W-5458971@
@P4CL:16616381@

+++ aura-impl/src/main/java/org/auraframework/impl/java/type/converter/LocalizedIntegerToStringConverter.java
@@ -53,7 +53,7 @@ public class LocalizedIntegerToStringConverter implements LocalizedConverter<Int
 
         try {
             Locale loc = locale.getNumberLocale();
-            return localizationService.formatNumber(value, loc);
+            return localizationService.formatNumber(value.intValue(), loc);
         } catch (Exception e) {
             return convert(value);
         }
